+++ server/src/main/java/org/allaymc/server/network/processor/impl/ingame/BlockPickRequestPacketProcessor.java
@@ -13,6 +13,8 @@ import org.cloudburstmc.protocol.bedrock.data.GameType;
 import org.cloudburstmc.protocol.bedrock.packet.BedrockPacketType;
 import org.cloudburstmc.protocol.bedrock.packet.BlockPickRequestPacket;
 
+import java.util.List;
+
 /**
  * @author Cool_Loong
  */
@@ -21,9 +23,9 @@ public class BlockPickRequestPacketProcessor extends PacketProcessor<BlockPickRe
     @Override
     public void handleSync(EntityPlayer player, BlockPickRequestPacket packet, long receiveTime) {
         var blockPos = MathUtils.CBVecToJOMLVec(packet.getBlockPosition());
-        // TODO: UserData
-        var addUserData = packet.isAddUserData();
-        if (!player.canReachBlock(blockPos) || player.getGameType() != GameType.CREATIVE) return;
+        if (!player.canReachBlock(blockPos) || player.getGameType() != GameType.CREATIVE) {
+            return;
+        }
 
         var block = player.getLocation().dimension().getBlockState(blockPos);
         if (block.getBlockType() == BlockTypes.AIR) {
@@ -34,16 +36,19 @@ public class BlockPickRequestPacketProcessor extends PacketProcessor<BlockPickRe
         var item = block.toItemStack();
         item.setCount(item.getItemData().maxStackSize());
 
-        var event = new PlayerBlockPickEvent(
-                player,
-                new BlockStateWithPos(block, new Position3i(blockPos, player.getDimension()), 0),
-                addUserData,
-                item
-        );
-        if (!event.call()) return;
+        var event = new PlayerBlockPickEvent(player, new BlockStateWithPos(block, new Position3i(blockPos, player.getDimension()), 0), packet.isAddUserData(), item);
+        if (!event.call()) {
+            return;
+        }
 
         item = event.getItemBlock();
-        addUserData = event.isAddUserData();
+        if (event.isIncludeBlockEntity()) {
+            var blockEntity = player.getDimension().getBlockEntity(blockPos);
+            if (blockEntity != null) {
+                item.setBlockEntityNBT(blockEntity.saveNBT());
+                item.setLore(List.of("+(DATA)"));
+            }
+        }
 
         var inventory = player.getContainer(FullContainerType.PLAYER_INVENTORY);
         // Foreach hot bar
@@ -63,10 +68,15 @@ public class BlockPickRequestPacketProcessor extends PacketProcessor<BlockPickRe
             }
         }
 
-        if (success) return;
+        if (success) {
+            return;
+        }
 
-        if (minEmptySlot != -1) inventory.setItemStack(minEmptySlot, item);
-        else inventory.setItemInHand(item); // Hot bar is full
+        if (minEmptySlot != -1) {
+            inventory.setItemStack(minEmptySlot, item);
+        } else {
+            inventory.setItemInHand(item); // Hot bar is full
+        }
     }
 
     @Override